fix(UI): selectAll button was still keeping alle items in memory

+++ UI/WebServerResources/js/Scheduler/CalendarListController.js
@@ -140,19 +140,20 @@
     }
 
     function confirmDeleteSelectedComponents() {
-      Dialog.confirm(l('Warning'),
-                     l('Are you sure you want to delete the selected components?'),
-                     { ok: l('Delete') })
-        .then(function() {
-          // User confirmed the deletion
-          var components = _.filter(Component['$' + vm.componentType], function(component) {
-            return component.selected;
+      var components = _.filter(Component['$' + vm.componentType], function(component) {
+        return component.selected;
+      });
+      if(components.length > 0)
+        Dialog.confirm(l('Warning'),
+                      l('Are you sure you want to delete the selected components?'),
+                      { ok: l('Delete') })
+          .then(function() {
+            // User confirmed the deletion
+            Calendar.$deleteComponents(components).then(function() {
+              vm.mode.multiple = 0;
+              $rootScope.$emit('calendars:list');
+            });
           });
-          Calendar.$deleteComponents(components).then(function() {
-            vm.mode.multiple = 0;
-            $rootScope.$emit('calendars:list');
-          });
-        });
     }
 
     function openEvent($event, event) {
